+++ tomli/_parser.py
@@ -4,7 +4,6 @@ from collections.abc import Iterable
 import string
 from types import MappingProxyType
 from typing import Any, BinaryIO, NamedTuple
-import warnings
 
 from tomli._re import (
     RE_DATETIME,
@@ -53,17 +52,7 @@ class TOMLDecodeError(ValueError):
 
 def load(fp: BinaryIO, *, parse_float: ParseFloat = float) -> dict[str, Any]:
     """Parse TOML from a binary file object."""
-    s_bytes = fp.read()
-    try:
-        s = s_bytes.decode()
-    except AttributeError:
-        warnings.warn(
-            "Text file object support is deprecated in favor of binary file objects."
-            ' Use `open("foo.toml", "rb")` to open the file in binary mode.',
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        s = s_bytes  # type: ignore[assignment]
+    s = fp.read().decode()
     return loads(s, parse_float=parse_float)